enable-ec_nistp_64_gcc_128
This requires a 64 bit system

Use on little endian platforms when GCC supports __uint128_t.
ECDH is about 2 to 4 times faster.
Not enabled by default for OpenSSL, because Configure can't determine it.
Enable it if your compiler defines __SIZEOF_INT128__,
the CPU is little endian and it tolerates unaligned data access.